% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_experiment.R
\name{add_constraint}
\alias{add_constraint}
\title{Add a constraint to the experimental factors}
\usage{
add_constraint(constraint)
}
\arguments{
\item{constraint}{a comparitive formula referencing treatment names}
}
\value{
invisible
}
\description{
Add a constraint to the experimental factors
}
\details{
Call is made using expressions, e.g. `experiment$add_constraint(x1 + x2 <= 14)`
}
