% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_experiment.R
\name{add_response}
\alias{add_response}
\title{Add a response to the experiment}
\usage{
add_response(name = paste0("y", length(self$responses) + 1),
  range = NULL, value = NULL, weight = 3)
}
\arguments{
\item{name}{a unique length 1 character vector giving the name of a response}

\item{range}{optional length 2 numeric vector giving boundaries on possible response levels}

\item{value}{function to compute the value of a particular response point,
the algorithm will maximize the response of this function}

\item{weight}{relative weighting factor used to prioritize some responses over others}
}
\value{
invisible
}
\description{
Add a response to the experiment
}
\details{
The value function must accept two positional arguments. The first is for the 'current' response
value while the second is for a vector of all input response values. These two inputs can then
be used however you like to return a numeric value. e.g. `function(y, hist){100 - y}`
}
