% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_experiment.R
\name{add_treatment}
\alias{add_treatment}
\title{Add a treatment to the experiment}
\usage{
add_treatment(name = paste0("x", length(self$treatments) + 1),
  boundaries = NULL)
}
\arguments{
\item{name}{a unique length 1 character vector giving the name of a treatment}

\item{boundaries}{optional length 2 numeric vector giving boundaries on possible treatment levels}
}
\value{
invisible
}
\description{
Add a treatment to the experiment
}
