% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utilities.R
\name{append_responses}
\alias{append_responses}
\title{Append experiment responses to simplex data frame}
\usage{
append_responses(simplex, responses)
}
\arguments{
\item{simplex}{data.frame of simplex}

\item{responses}{vector of response names}
}
\value{
data.frame of simplex with responses appended
}
\description{
Append experiment responses to simplex data frame
}
