% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_experiment.R
\name{check_response_range}
\alias{check_response_range}
\title{Check if provided response is in stated range}
\usage{
check_response_range(val, responses, r)
}
\arguments{
\item{val}{The numeric response value provided}

\item{responses}{The list of responses from self}

\item{r}{The name of the response being evaluated}
}
\value{
TRUE if it doesn't error out
}
\description{
Check if provided response is in stated range
}
