% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_experiment_utilities.R
\name{determine_cohypoplanarity}
\alias{determine_cohypoplanarity}
\title{Determine if simplex is degenerate (cohypolanar)}
\usage{
determine_cohypoplanarity(simplex)
}
\arguments{
\item{simplex}{data frame with k+1 rows and k columns}
}
\value{
logical indicating if simplex is cohypoplanar
}
\description{
Determine if simplex is degenerate (cohypolanar)
}
