% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_experiment.R
\name{drop_constraint}
\alias{drop_constraint}
\title{Drop a constraint from the experiment}
\usage{
drop_constraint(id)
}
\arguments{
\item{id}{the id of the constraint to drop}
}
\value{
invisible
}
\description{
Drop a constraint from the experiment
}
