% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_experiment.R
\name{drop_response}
\alias{drop_response}
\title{Drop a response from the experiment}
\usage{
drop_response(name)
}
\arguments{
\item{name}{a unique length 1 character vector giving the name of the response to drop}
}
\value{
invisible
}
\description{
Drop a response from the experiment
}
