% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_experiment.R
\name{drop_treatment}
\alias{drop_treatment}
\title{Drop a treatment from the experiment}
\usage{
drop_treatment(name)
}
\arguments{
\item{name}{a unique length 1 character vector giving the name of the treatment to drop}
}
\value{
invisible
}
\description{
Drop a treatment from the experiment
}
