% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_experiment_class.R
\docType{class}
\name{experiment}
\alias{experiment}
\title{Class for managing sequential simplex experiments}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
experiment
}
\description{
Class for managing sequential simplex experiments
}
\keyword{datasets}
