% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_experiment.R
\name{generate_initial_simplex}
\alias{generate_initial_simplex}
\title{Generate initial simplex}
\usage{
generate_initial_simplex(method = "manual", data = NULL)
}
\arguments{
\item{method}{Single character value determining the method of initial simplex calculation}

\item{data}{Data Frame with relevant information, see details}
}
\value{
invisible
}
\description{
Generate initial simplex
}
\details{
The parameter 'method' can take three values, "manual", "corner" or "tilted". Each coresponds to a different method
for determining the initial simplex.

*manual*
When the initial simplex is manually determined, the parameter 'data' must contain a data frame with one column
per treatment (sharing the same name as the treatment) and the k+1 initial vertices.

*corner* and *tilted*
These two methods for automatically generating a starting simplex require that the parameter 'data' contain a 3
column data frame with 1 row per treatment. The first column contains treatment names, the second contains the
starting coordinate for the treatment and the third contains the step size for the treatment.

The initial simplex has as the first vertex the combination of all starting coordinates. Each sequential vertex is
produced by adding some amount of the step size to one or more of the other treatments.
}
