% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utilities.R
\name{\%between\%}
\alias{\%between\%}
\title{Check if values are between two other values}
\usage{
x \%between\% r
}
\arguments{
\item{x}{the value(s) to be checked (left hand side)}

\item{r}{range to check if is between}
}
\value{
logical indicating if x is between min/max of r
}
\description{
Check if values are between two other values
}
\details{
This is an implementation specific to this package as it returns TRUE if r is null.
Not exported as it's not intended for broader use
}
