% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utilities.R
\name{init}
\alias{init}
\title{Initialize the R6 Class}
\usage{
init(k = NA_integer_)
}
\arguments{
\item{k}{dimensionality of space being experimented}
}
\value{
invisible
}
\description{
Initialize the R6 Class
}
\examples{
experiment$new(k = 3)
}
