% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utilities.R
\name{satisfied_constraints}
\alias{satisfied_constraints}
\title{Check if all experimental constraints are met by the simplex}
\usage{
satisfied_constraints(simplex, constraints)
}
\arguments{
\item{simplex}{data.frame containing the simplex to be checked}

\item{constraints}{list of treatment constraints}
}
\value{
logical indicating if all constraints are met
}
\description{
Check if all experimental constraints are met by the simplex
}
