% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_experiment.R
\name{submit_responses}
\alias{submit_responses}
\title{Method to enable user to submit responses to experiment interactively}
\usage{
submit_responses(data = NULL)
}
\arguments{
\item{data}{optional input to pass in responses}
}
\value{
invisible
}
\description{
Method to enable user to submit responses to experiment interactively
}
\details{
If data is manually passed in, it must be a data.frame with one column per resonse
plus one column named "Vertex_ID" which must match
}
