% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utilities.R
\name{valid_simplex}
\alias{valid_simplex}
\title{Check if simplex is valid}
\usage{
valid_simplex(simplex, names, boundaries, constraints)
}
\arguments{
\item{simplex}{data.frame of simplex to be checked}

\item{names}{vector of treatment names}

\item{boundaries}{list of treatment boundaries}

\item{constraints}{list of treatment constraints}
}
\value{
logical indicating if simplex is valid
}
\description{
Check if simplex is valid
}
\details{
PRIVATE
}
