% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_utilities.R
\name{within_boundaries}
\alias{within_boundaries}
\title{Check if all simplex values are within boundaries}
\usage{
within_boundaries(simplex, names, boundaries)
}
\arguments{
\item{simplex}{data.frame containing the simplex to be checked}

\item{names}{vector of treatment names}

\item{boundaries}{list of treatment boundaries}
}
\value{
logical indicating if all values are within boundaries
}
\description{
Check if all simplex values are within boundaries
}
